backend/test_openai.py
import os

from dotenv import load_dotenv
from openai import OpenAI

load_dotenv()

try:
    client = OpenAI(api_key=os.getenv("OPENAI_API_KEY"))
    print("✅ OpenAI client created successfully")
    print(f"API Key loaded: {'Yes' if os.getenv('OPENAI_API_KEY') else 'No'}")

    # Test a simple API call
    response = client.chat.completions.create(
        model="gpt-3.5-turbo",
        messages=[{"role": "user", "content": "Say hello"}],
        max_tokens=10,
    )
    print("✅ OpenAI API call successful")
    print(f"Response: {response.choices[0].message.content}")

except Exception as e:
    print(f"❌ Error: {e}")
