backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.11.7
pytest==7.4.3
httpx==0.25.2
black==23.12.1
flake8==6.1.0
isort==5.13.2